+++ packages/core/test/Form_test.js
@@ -252,6 +252,30 @@ describeRepeated("Form common", createFormComponent => {
     });
   });
 
+  describe("Option idSeparator", function() {
+    it("should change the rendered ids", function() {
+      const schema = {
+        type: "object",
+        title: "root object",
+        required: ["foo"],
+        properties: {
+          count: {
+            type: "number",
+          },
+        },
+      };
+      const comp = renderIntoDocument(<Form schema={schema} idSeparator="." />);
+      const node = findDOMNode(comp);
+      const inputs = node.querySelectorAll("input");
+      const ids = [];
+      for (var i = 0, len = inputs.length; i < len; i++) {
+        const input = inputs[i];
+        ids.push(input.getAttribute("id"));
+      }
+      expect(ids).to.eql(["root.count"]);
+    });
+  });
+
   describe("Custom field template", () => {
     const schema = {
       type: "object",