+++ CHANGELOG.md
@@ -40,6 +40,9 @@ should change the heading of the (upcoming) version to include a major version b
 - Differentiated the material-ui 4 and 5 themes
 - Added chakra-ui theme
 
+## @rjsf/core
+- introduce `idSeparator` property (https://github.com/rjsf-team/react-jsonschema-form/pull/2628)
+
 # v3.3.0
 
 ## @rjsf/semantic-ui